Add files via upload
This is a simple project that uses SQL to do some data exploration. I will use this to create visualizations in my next project using Tableau. The data was taken from https://ourworldindata.org/covid-deaths, which I then fine tuned to create tables that examine the deaths and vaccinations associated with covid, as well as a list of queries I used to explore and extract said data.